// app/src/generated-api/models/CharacterItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Underworld
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EquipLocationTag,
    EquipLocationTagFromJSON,
    EquipLocationTagFromJSONTyped,
    EquipLocationTagToJSON,
} from './EquipLocationTag';
import {
    Item,
    ItemFromJSON,
    ItemFromJSONTyped,
    ItemToJSON,
} from './Item';

/**
 * 
 * @export
 * @interface CharacterItem
 */
export interface CharacterItem {
    /**
     * 
     * @type {Item}
     * @memberof CharacterItem
     */
    item: Item;
    /**
     * 
     * @type {boolean}
     * @memberof CharacterItem
     */
    is_hidden?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CharacterItem
     */
    knows_equipped_location: boolean;
    /**
     * 
     * @type {EquipLocationTag}
     * @memberof CharacterItem
     */
    equipped_location?: EquipLocationTag;
    /**
     * 
     * @type {boolean}
     * @memberof CharacterItem
     */
    is_multiple: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CharacterItem
     */
    at_the_ready: boolean;
}

export function CharacterItemFromJSON(json: any): CharacterItem {
    return CharacterItemFromJSONTyped(json, false);
}

export function CharacterItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): CharacterItem {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'item': ItemFromJSON(json['item']),
        'is_hidden': !exists(json, 'is_hidden') ? undefined : json['is_hidden'],
        'knows_equipped_location': json['knows_equipped_location'],
        'equipped_location': !exists(json, 'equipped_location') ? undefined : EquipLocationTagFromJSON(json['equipped_location']),
        'is_multiple': json['is_multiple'],
        'at_the_ready': json['at_the_ready'],
    };
}

export function CharacterItemToJSON(value?: CharacterItem | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'item': ItemToJSON(value.item),
        'is_hidden': value.is_hidden,
        'knows_equipped_location': value.knows_equipped_location,
        'equipped_location': EquipLocationTagToJSON(value.equipped_location),
        'is_multiple': value.is_multiple,
        'at_the_ready': value.at_the_ready,
    };
}